Close navigation popup for router links
When nagivating to a on-hanger page, like authors or the resource
guidelines, the nagivation popup is not closed automatically all the
time as the vue-router does not properly focus the new page elements
after routing.

To improve user experience, this commit simply closes the popup navbar
once an on-site link is clicked.
This behaviour is not needed for external links as those leave the
hangar site.