Configure vue-router to scroll to top
Up till now, the vue-router did not have any scroll behaviour, leading
to no scrolling on page changes. This becomes particularly obvious when
opening a project from bottom of the main project list, as the project
page is scrolled down to a good part of the README.md.

To fix this, this commit configures the routers scrollBehaviour
following the vuejs documentation to create the most native scrolling
behaviour possible, scrolling to any potentially saved position when
using the browsers back button or scrolling to the top of the page
otherwise.

See: https://router.vuejs.org/guide/advanced/scroll-behavior.html