Disable next button on incorrect new version data
When uploading a new version of a project to hangar, the next buttons in
the steps screens 'Artifact' and 'Basic Info' are not greyed out even if
invalid data is currently present in the form.

Specifically, the next button in the 'Artifact' screen is visible usable
even if the current url is not a valid url.
Additionally, the next button in the 'Basic Info' step is visible
usable, (and logically usable), even without a single platform selected,
leading to project creation errors down the line.

Beyond that, this commit also fixes a slight oversite in which the
description was able to simply pass null as the raw prop of the markdown
editor instead of an empty string, resulting in a client side error.