Prevent Steps#next on final version creation step
Different than the project creation usage of the Steps component, which
adds a proper final creation button, the version creation page simply
uses the beforeNext function of the final step to create the version and
route the client to the version overview.

Previously however, this logic would still instruct the Steps component
to move forward to the next step if the creation succeeded. However this
raises an index out of bounds exception, as this logic runs on the last
step.

To prevent this, this commit always disallows movement to the next,
non-existing, step and simply lets the router call in the beforeNext
method take care of routing.