Create tokens with same-site policy lax
Both the HangarAuth and HangarAuth_REFRESH cookies were set with
same-site policy 'strict', preventing the browser from sending them when
opening a hangar page from a third party href, such as a plain link in a
github gist.

This prevents the server side rendering from picking up the token or
refresh token, leading the the server side rendering logic to assume an
unauthorized user is attempting to view a page eventho the user is
actually logged in.
This directly leads to hydration errors as the SSR renders the page for
a non-logged in user while the client side hydration properly detects
the logged in user and renders a potentially completely different page.