Sanitize trailing slashes from project page urls
While looking up a project page, the frontend may be passing in a path
that contains a trailing slash. For example
'page/user/project/pages/ExamplePage/'.

This, prior to this commit, yielded a 404 not found response which in
turn messed with SSR.

To prevent this, the hangar frontend now actively sanitizes the url
before requesting the page from the backend.