Remove invalid versions when filtering platforms
When users filter the available projects by platform and versions,
switching the platform, prior to this commit, does not remove versions
no available on the platform selected.

This leads to empty return queries, e.g. when a user first selects
version 1.19 while filtering for the paper platform and then switching
to the velocity view, which simply does not have a 1.19 version.

This commit fixes this faulty behaviour by removing versions that are
not available in the platform filtered by. Versions that match, which is
usually the case when switching between paper and waterfall, are kept as
a general improvement to user experience compared to always clearing the
versions selected.