Update fetched project page tree on mutation
When a new project page is added or removed to or from the project, the
frontend does not completely reload the project, which is fine. However
this means that, until a complete page reload is invoked, the pages in
the project page tree will not reflect the new page structure of the
project.

To fix this, this commit defines a new provider on the root project page
that provides a simple update function taking in the new project page
array.

Child components way down the tree, specifically the NewPageModel and
the ProjectPageMarkdown component then inject said update function and
feed it the new page layout fetched from the backend on project page
index mutation.

Resolves: #713