Use the proper config for page name validation
Prior to this commit, hangar would use the project name configuration to
check the page name. This completely ignored the existing configuration
for page names, which already permits spaces and has existing rules
regarding max and min length.

This commit now properly uses said configuration feature when checking
the name of a page that is to be created.