Reset the uploaded image when it's too big
When uploading a new avatar, if the image is too big it will notify the user, however it does not reset the uploaded image making it appear like you can continue to upload it. Now it will instead reset the image, making it obvious that the image cannot be uploaded.