Fix inconsistency in the keyword tag entry box
This fixes #1183. Due to the way the maxLength on the input type works, if we are at the 15 character limit for a keyword it will not allow us to add a space. By handling the max length ourselves, we're able to more specifically allow spaces / commas once we've hit the maxLength. This also means not using the v-model approach to gain more customizability. There were also places where the tag-maxlength were not added to the Input element, causing projects to have tags that are longer than 15 characters.